fix(ui5-proxy-middleware): support usage inside connect servers
When using the dev-approuter (https://www.npmjs.com/package/dev-approuter)
to run UI5 and CAP applications inside the @sap/router from a monorepo
then the server is using connect and not express. In those scenarios, similar
like with livereload only the basic http response object from Node.js is
available and the status function is not available which leads to an
INTERNAL SERVER ERROR. By using the writeHead, write and end function of
the response object the function is agnostic to the used server and works
in both scenarios.