feat(icons): add new Manage icon (#2639)
* feat(icons): add new Manage icon

Introduce a new 'Manage' icon to the UIIcons component. This enhancement expands the icon library, allowing for more diversified icon usage in UI components.

* feat: add 'Manage' icon to UIIcons component

Introduce a new 'Manage' icon within the UIIcons component to expand the icon library. This enhancement allows developers to utilize a wider variety of icons, improving the visual diversity and usability of UI components.